pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "coding_agent"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.43",
    "llmlib-amq==0.0.7",
    "pydantic>=2.9.2",
    "rich>=13.9.2",
    "typer>=0.12.5",
]
dynamic = ["version"]

[tool.uv]
dev-dependencies = [
    "flit>=3.9.0",
    "invoke>=2.2.0",
    "isort>=5.13.2",
    "mypy>=1.12.0",
    "pylint>=3.3.1",
    "pytest-xdist>=3.6.1",
    "pytest>=8.3.3",
    "ruff>=0.7.0",
]

[tool.flit.module]
name = "coding_agent"

#[tool.flit.sdist]
#include = ["LICENSE"]


[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring",
    "broad-exception-caught",
    "too-few-public-methods"
    # Add more Pylint rules to disable here
]

[tool.ruff]
#select = ["E", "F", "I"]
#ignore = [
#    "E501",  # line too long
#     Add more Ruff rules to disable here
#]

#[tool.ruff.per-file-ignores]
#"__init__.py" = ["F401"]  # unused import
